Re-implementation of the produce_matrix function
The goal of this re-implementation of the produce_matrix function is to
allow easier customization of the scores to calculate for every
positions in the matrix.

This also fixed a bug where invalid values were calculated for matrix
with different GRangesList for x and y. This bug was caused by the fact
that i,j and j,i positions in the matrix was calculated at the same
time to save time. But this was leading to invalid results in some
case.

Thus, the new version takes a bit longer to calculate since every
position in the matrix is calculated independently, but it is easier to
implement a new scoring function.

To reduce the time required to compute the values of the matrix, it is
now possible to use multiple cores (with the cores parameter in
chipcompare initilize (new) function).

Signed-off-by: Charles Joly Beauparlant <Charles.Joly-Beauparlant@crchul.ulaval.ca>