Added a function to calculate p-value
In order to determine if an observed overlap is significant, we have
added the possibility to calculate the p-value for every elements in the
matrix.

The range of expected overlap is simulated using bootstrap with a
baseline (a set of observed regions with as many datasets from the same
cell type as possible). The observed overlap is then compared to the
boostrapped values to obtain the p-value.

Signed-off-by: Charles Joly Beauparlant <Charles.Joly-Beauparlant@crchul.ulaval.ca>